Title: 최소 통과 시간, Time: 68ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240720/최소 통과 시간/minimum-transit-time.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-long long n, m;
+int n, m;
 long long arr[100001], ans = 21e13;
 
 void ps(long long start, long long end)
@@ -14,7 +14,7 @@ void ps(long long start, long long end)
         mid = (start + end) / 2;
 
         long long cnt = 0;
-        for(long long i=0;i<m;i++) cnt += (mid / arr[i]);
+        for(int i=0;i<m;i++) cnt += (mid / arr[i]);
 
         if (cnt >= n)
         {
@@ -32,7 +32,7 @@ int main() {
 
     long long mini = 21e13, maxi = -21e13;
 
-    for(long long i=0;i<m;i++)
+    for(int i=0;i<m;i++)
     {
         cin >> arr[i];
